Hotfix: Resolve conflicts (#13)
* Fix some std::clamp() calls: max value cannot be less than min value

* Improve flip_image() performance (x10 from old impl, x2 from new slow impl)

The old impl was using get/put_pixel(), the new slow one is using the
get/put_pixel_fast(), and the new default flip_image() is using just
raw pointers.

Added some utilites like random_image() for testing purposes, and
DOC_DISPATCH_BY_COLOR_MODE() macros to avoid switch/case for each
color mode. In a future these might use generic lambdas.

* Update actions to use new node16 as GitHub is deprecating node12

* Update laf module

* [lua] Add traceback info when a doc object is deleted and we try to use it

A reference to a doc object (Sprite, Layer, Cel, etc.) is done through
its ID in the scripting engine, when we try to access it from a script
that element might be already deleted. Previously we displayed the a
message like "Using a nil 'Cel' object". With this change we show the
traceback and a "Tried to access a deleted 'Cel'" message.

* Minor fix in doc_range_tests in the internal image used

---------

Co-authored-by: David Capello <david@igara.com>